MiniGPT-v2/requirement_minigpt4.txt
absl-py==2.1.0
accelerate==0.28.0
aiofiles==23.2.1
aiohttp==3.9.3
aiosignal==1.3.1
allennlp==2.10.1
altair==5.3.0
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==4.3.0
async-timeout==4.0.3
attrs==23.2.0
base58==2.1.1
bitsandbytes==0.43.0
blis==0.7.11
boto3==1.34.77
botocore==1.34.77
braceexpand==0.1.7
cached-path==1.1.6
cachetools==5.3.3
catalogue==2.0.10
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
cmake==3.29.0.1
commonmark==0.9.1
confection==0.1.4
contextlib2==21.6.0
contourpy==1.2.1
cycler==0.12.1
cymem==2.0.8
decord==0.6.0
dill==0.3.8
docker-pycreds==0.4.0
exceptiongroup==1.2.0
fairscale==0.4.6
fastapi==0.110.1
ffmpy==0.3.2
filelock==3.7.1
fonttools==4.50.0
frozenlist==1.4.1
fschat==0.2.34
fsspec==2024.3.1
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.18.0
google-auth==2.29.0
google-cloud-core==2.4.1
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
gradio==3.50.0
gradio_client==0.6.1
h11==0.14.0
h5py==3.10.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.22.2
idna==3.6
imageio==2.34.0
importlib_resources==6.4.0
iniconfig==2.0.0
iopath==0.1.10
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.3.2
jsonnet==0.20.0
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
langcodes==3.3.0
lazy_loader==0.3
lit==18.1.2
lmdb==1.4.1
markdown-it-py==3.0.0
markdown2==2.4.13
MarkupSafe==2.1.5
matplotlib==3.8.4
mdurl==0.1.2
ml-collections==0.1.1
more-itertools==10.2.0
mpmath==1.3.0
multidict==6.0.5
murmurhash==1.0.10
networkx==3.2.1
nh3==0.2.17
nltk==3.8.1
numpy==1.26.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu11==10.2.10.91
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu11==11.7.4.91
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu11==2.14.3
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu11==11.7.91
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
opencv-python==4.9.0.80
orjson==3.10.0
packaging==24.0
pandas==2.2.1
pathlib_abc==0.1.1
pathtools==0.1.2
pathy==0.11.0
peft==0.2.0
pillow==10.3.0
pluggy==1.4.0
portalocker==2.8.2
preshed==3.0.9
progressbar2==4.4.2
promise==2.3
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==3.20.3
psutil==5.9.8
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.6.4
pydantic_core==2.16.3
pydub==0.25.1
Pygments==2.17.2
pyparsing==3.1.2
pytest==8.1.1
python-dateutil==2.9.0.post0
python-multipart==0.0.9
python-utils==3.8.2
pytz==2024.1
PyYAML==6.0.1
referencing==0.34.0
regex==2023.12.25
requests==2.31.0
rich==12.6.0
rpds-py==0.18.0
rsa==4.9
ruff==0.3.5
s3transfer==0.10.1
sacremoses==0.1.1
safetensors==0.4.2
scikit-image==0.22.0
scikit-learn==1.4.1.post1
scipy==1.13.0
seaborn==0.13.2
semantic-version==2.10.0
sentencepiece==0.2.0
sentry-sdk==1.44.1
setproctitle==1.3.3
shellingham==1.5.4
shortuuid==1.0.13
six==1.16.0
smart-open==6.4.0
smmap==5.0.1
sniffio==1.3.1
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
starlette==0.37.2
svgwrite==1.4.3
sympy==1.12
tensorboardX==2.6.2.2
termcolor==1.1.0
thinc==8.2.3
threadpoolctl==3.4.0
tifffile==2024.2.12
tiktoken==0.6.0
timm==0.9.16
tokenizers==0.12.1
tomli==2.0.1
tomlkit==0.12.0
toolz==0.12.1
torch==2.2.2
torchvision==0.13.1
tqdm==4.66.2
traitlets==5.14.2
transformers==4.30.0
triton==2.2.0
typer==0.11.1
typer-cli==0.12.0
typer-slim==0.12.0
typing_extensions==4.10.0
tzdata==2024.1
urllib3==1.26.18
uvicorn==0.29.0
visual-genome==1.1.1
wandb==0.12.21
wasabi==0.10.1
wavedrom==2.0.3.post3
wcwidth==0.2.13
weasel==0.3.4
webdataset==0.2.86
websockets==11.0.3
yarl==1.9.4
zipp==3.18.1
